pyproject.toml
[build-system]
requires = ["setuptools >= 80.0.0", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cricket-warehouse"
dynamic = ["version"]
dependencies = [
    "apache-airflow==2.9.3",
    "apache-airflow-providers-postgres>=6.1.3",
    "dbt-core>=1.11.7",
    "dbt-postgres>=1.10.0",
    "platformdirs>=4.9.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "typer>=0.24.1",
]
requires-python = ">=3.10"
authors = [
	{name="Siddhant Sharma", email="sharmasiddhant@protonmail.com" },
]
description = "Build a cricket data warehouse for analytics."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
keywords = []


[tool.setuptools_scm]
local_scheme="no-local-version"
fallback_version = "0.0.0"

[project.scripts]
cricwh = "cricketwarehouse.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "dbt-core>=1.11.7",
    "dbt-postgres>=1.10.0",
    "ipython>=8.38.0",
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.3.260113",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
    "types-psycopg2>=2.9.21.20251012",
]

[project.urls]
Repository = "https://github.com/shsiddhant/cricket-warehouse"
Issues = "https://github.com/shsiddhant/cricket-warehouse/issues"
